requirements.txt
alembic==1.8.1
asttokens==2.1.0
autopep8==2.0.0
backcall==0.2.0
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.0
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future==0.18.2
google-auth==2.13.0
google-auth-oauthlib==0.6.0
greenlet==1.1.3.post0
gspread==5.6.2
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pywin32==304
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.42
stack-data==0.6.0
tomli==2.0.1
torch==1.12.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
viberbot==1.0.12
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
